IAT-46:  Cleaned up the rebase from the develop branch regarding IAT-458 model refactoring.
