+++ src/main/java/org/opentestsystem/ap/ims/model/ItemHistory.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.opentestsystem.ap.ims.entity;
+package org.opentestsystem.ap.ims.model;
 
 import java.util.Date;
