+++ src/main/java/org/opentestsystem/ap/ims/model/ItemMapper.java
@@ -23,6 +23,7 @@ import javax.xml.bind.JAXBElement;
 
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.gitlab4j.api.models.Commit;
 import org.opentestsystem.saaif.item.ItemRelease;
 import org.opentestsystem.saaif.item.ObjectFactory;
 import org.springframework.stereotype.Component;
@@ -49,6 +50,20 @@ import static org.opentestsystem.ap.ims.model.ItemConstants.ItemVersion.ITEM_VER
 public class ItemMapper {
 
     private static final ObjectFactory OBJECT_FACTORY = new ObjectFactory();
+
+    // ------------------------------------------------------------------------
+
+    public static ItemHistory mapToEntityItemHistory(final Commit gitlabCommit) {
+        final ItemHistory itemHistory = ItemHistory.ItemHistoryBuilder
+            .anItemHistory()
+            .changedBy(gitlabCommit.getCommitterName())
+            .changedOn(gitlabCommit.getCommittedDate())
+            .changedByEmail(gitlabCommit.getCommitterEmail())
+            .message(gitlabCommit.getMessage())
+            .build();
+        return itemHistory;
+    }
+
 
     // ------------------------------------------------------------------------
