+++ src/main/java/org/opentestsystem/ap/ims/repository/ItemRepository.java
@@ -21,13 +21,14 @@ import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.gitlab4j.api.models.Commit;
-import org.gitlab4j.api.models.Project;
 import org.opentestsystem.ap.ims.client.GitClient;
 import org.opentestsystem.ap.ims.client.GitClientFactory;
 import org.opentestsystem.ap.ims.client.GitlabClient;
 import org.opentestsystem.ap.ims.model.Item;
 import org.opentestsystem.ap.ims.model.ItemBankUser;
 import org.opentestsystem.ap.ims.model.ItemFactory;
+import org.opentestsystem.ap.ims.model.ItemHistory;
+import org.opentestsystem.ap.ims.model.ItemMapper;
 import org.opentestsystem.ap.ims.util.ItemIdGenerator;
 import org.opentestsystem.ap.ims.util.ResourceNotFoundException;
 import org.opentestsystem.ap.ims.util.ValidationException;
@@ -71,8 +72,8 @@ public class ItemRepository {
     public List<ItemHistory> findItemHistory(final String itemId) {
         log.debug("find item history for {}", itemId);
         final List<Commit> commits = gitlabClient.lookupItemHistory(itemId);
-        final List<ItemHistory> itemHistoryList = commits.stream().map(commit -> itemMapper.mapToEntityItemHistory
-            (commit)).collect(Collectors.toList());
+        final List<ItemHistory> itemHistoryList = commits.stream().map(ItemMapper::mapToEntityItemHistory).collect
+            (Collectors.toList());
         return itemHistoryList;
     }
