+++ src/main/java/org/opentestsystem/ap/ims/rest/v1/ItemApi.java
@@ -20,8 +20,9 @@ import java.util.List;
 import io.swagger.annotations.Api;
 import lombok.extern.slf4j.Slf4j;
 import org.opentestsystem.ap.ims.model.Item;
-import org.opentestsystem.ap.ims.rest.ItemRequestWrapper;
+import org.opentestsystem.ap.ims.model.ItemHistory;
 import org.opentestsystem.ap.ims.rest.CreateNewItemRequest;
+import org.opentestsystem.ap.ims.rest.ItemRequestWrapper;
 import org.opentestsystem.ap.ims.service.ItemService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -33,8 +34,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import static org.opentestsystem.ap.ims.util.AppConstants.API_V1;
-
-import org.opentestsystem.ap.ims.entity.ItemHistory;
 
 /**
  * Operations on items in the item bank.
