+++ src/main/java/org/opentestsystem/ap/ims/service/ItemService.java
@@ -26,7 +26,7 @@ import org.opentestsystem.ap.ims.util.SecurityUtil;
 import org.opentestsystem.ap.ims.util.ValidationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.opentestsystem.ap.ims.entity.ItemHistory;
+import org.opentestsystem.ap.ims.model.ItemHistory;
 
 /**
  * The item management service.
