+++ src/test/java/org/opentestsystem/ap/ims/model/ItemMapperTest.java
@@ -16,17 +16,21 @@
 package org.opentestsystem.ap.ims.model;
 
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.xml.bind.JAXBElement;
 
 import org.assertj.core.util.Lists;
+import org.gitlab4j.api.models.Commit;
 import org.junit.Test;
 import org.opentestsystem.saaif.item.ItemRelease;
 import org.opentestsystem.saaif.item.ObjectFactory;
 
 import static org.apache.commons.lang3.StringUtils.EMPTY;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 import static org.opentestsystem.ap.ims.model.ItemConstants.EXEMPLAR_PURPOSE;
 import static org.opentestsystem.ap.ims.model.ItemConstants.EXEMPLAR_RESPONSE;
 import static org.opentestsystem.ap.ims.model.ItemConstants.EXEMPLAR_SCOREPOINT;
@@ -42,6 +46,29 @@ public class ItemMapperTest {
 
     private final ObjectFactory OBJECT_FACTORY = new ObjectFactory();
 
+
+    @Test
+    public void itShouldMapToEntityItemHistory() {
+        final String committerEmail = "test@fake.com";
+        final String committerName = "John Doe";
+        final Date committedDate = new Date();
+        final String message = "this is a commit message";
+
+        final Commit mockCommit = mock(Commit.class);
+
+        when(mockCommit.getCommitterName()).thenReturn(committerName);
+        when(mockCommit.getCommitterEmail()).thenReturn(committerEmail);
+        when(mockCommit.getCommittedDate()).thenReturn(committedDate);
+        when(mockCommit.getMessage()).thenReturn(message);
+
+        final ItemHistory itemHistory = ItemMapper.mapToEntityItemHistory(mockCommit);
+
+        assertThat(itemHistory.getChangedBy()).isEqualTo(committerName);
+        assertThat(itemHistory.getChangedByEmail()).isEqualTo(committerEmail);
+        assertThat(itemHistory.getMessage()).isEqualTo(message);
+        assertThat(itemHistory.getChangedOn()).isEqualTo(committedDate);
+
+    }
 
     @Test
     public void itShouldNewSaaifItemAttributeList() {
