+++ src/test/java/org/opentestsystem/ap/ims/repository/ItemRepositoryTest.java
@@ -16,6 +16,7 @@
 package org.opentestsystem.ap.ims.repository;
 
 import java.io.IOException;
+import java.util.Date;
 import java.util.List;
 
 import org.assertj.core.util.Lists;
@@ -30,8 +31,9 @@ import org.mockito.runners.MockitoJUnitRunner;
 import org.opentestsystem.ap.ims.client.GitClient;
 import org.opentestsystem.ap.ims.client.GitClientFactory;
 import org.opentestsystem.ap.ims.client.GitlabClient;
-import org.opentestsystem.ap.ims.entity.Item;
-import org.opentestsystem.ap.ims.entity.ItemHistory;
+import org.opentestsystem.ap.ims.model.Item;
+import org.opentestsystem.ap.ims.model.ItemFactory;
+import org.opentestsystem.ap.ims.model.ItemHistory;
 import org.opentestsystem.ap.ims.util.IMSTestUtil;
 import org.opentestsystem.ap.ims.util.ItemIdGenerator;
 import org.opentestsystem.ap.ims.util.ResourceNotFoundException;
@@ -75,8 +77,6 @@ public class ItemRepositoryTest {
     @Mock
     private ItemFactory mockItemFactory;
 
-    @Mock
-    private ItemMapper mockItemMapper;
 
     @Mock
     Item mockItem = mock(Item.class);
@@ -106,18 +106,27 @@ public class ItemRepositoryTest {
 
     @Test
     public void itShouldFindItemHistory() {
+        final Date committedDate = new Date();
         final Commit mockCommit = mock(Commit.class);
+
         final ItemHistory mockItemHistory = mock(ItemHistory.class);
+
+        when(mockCommit.getCommitterName()).thenReturn("name");
+        when(mockCommit.getCommitterEmail()).thenReturn("email");
+        when(mockCommit.getCommittedDate()).thenReturn(committedDate);
+        when(mockCommit.getMessage()).thenReturn("message");
 
         when(mockGitlabClient.lookupItemHistory(ITEM_ID)).thenReturn(Lists.newArrayList(mockCommit));
 
-        when(mockItemMapper.mapToEntityItemHistory(mockCommit)).thenReturn(mockItemHistory);
-
         final List<ItemHistory> itemHistoryList = repository.findItemHistory(ITEM_ID);
 
-        verify(mockItemMapper, times(1)).mapToEntityItemHistory(mockCommit);
-
         assertThat(itemHistoryList).hasSize(1);
+
+        final ItemHistory itemHistory = itemHistoryList.get(0);
+        assertThat(itemHistory.getChangedBy()).isEqualTo("name");
+        assertThat(itemHistory.getChangedByEmail()).isEqualTo("email");
+        assertThat(itemHistory.getChangedOn()).isEqualTo(committedDate);
+        assertThat(itemHistory.getMessage()).isEqualTo("message");
     }
 
     @Test(expected = ValidationException.class)
