+++ src/test/java/org/opentestsystem/ap/ims/service/ItemServiceIT.java
@@ -16,6 +16,8 @@
 
 package org.opentestsystem.ap.ims.service;
 
+import java.util.List;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.gitlab4j.api.models.Commit;
@@ -28,10 +30,10 @@ import org.junit.runner.RunWith;
 import org.opentestsystem.ap.ims.client.GitClient;
 import org.opentestsystem.ap.ims.client.GitClientFactory;
 import org.opentestsystem.ap.ims.client.GitlabClient;
-import org.opentestsystem.ap.ims.entity.Attribute;
-import org.opentestsystem.ap.ims.entity.Item;
-import org.opentestsystem.ap.ims.entity.ItemBankUser;
-import org.opentestsystem.ap.ims.entity.ItemHistory;
+import org.opentestsystem.ap.ims.model.Item;
+import org.opentestsystem.ap.ims.model.ItemBankUser;
+import org.opentestsystem.ap.ims.model.ItemHistory;
+import org.opentestsystem.ap.ims.model.SaItem;
 import org.opentestsystem.ap.ims.repository.ItemRepository;
 import org.opentestsystem.ap.ims.util.IMSTestUtil;
 import org.springframework.beans.factory.annotation.Autowired;
